+++ tests/gateways/disk/test_delete.py
@@ -1,20 +1,19 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-import os
-import pytest
 from errno import ENOENT
-from os.path import join, isdir, islink, lexists, isfile
-from conda.base.context import context
-from conda.gateways.disk.delete import rm_rf, move_to_trash
+import os
+from os.path import isdir, isfile, islink, join, lexists
+
+import pytest
+
+from conda.compat import TemporaryDirectory
+from conda.gateways.disk.create import create_link
+from conda.gateways.disk.delete import move_to_trash, rm_rf
+from conda.gateways.disk.link import islink, symlink
 from conda.gateways.disk.update import touch
-from conda.gateways.disk.link import islink
-from .test_permissions import tempdir, _try_open, _make_read_only
-from conda.utils import on_win
-
-
-def can_not_symlink():
-    return on_win and context.default_python[0] == '2'
+from conda.models.enums import LinkType
+from .test_permissions import _make_read_only, _try_open, tempdir
 
 
 def _write_file(path, content):
@@ -62,13 +61,12 @@ def test_remove_dir():
         assert not lexists(test_path)
 
 
-@pytest.mark.skipif(can_not_symlink(), reason="symlink function not available")
 def test_remove_link_to_file():
     with tempdir() as td:
         dst_link = join(td, "test_link")
         src_file = join(td, "test_file")
         _write_file(src_file, "welcome to the ministry of silly walks")
-        os.symlink(src_file, dst_link)
+        symlink(src_file, dst_link)
         assert isfile(src_file)
         assert not islink(src_file)
         assert islink(dst_link)
@@ -80,13 +78,12 @@ def test_remove_link_to_file():
         assert not lexists(dst_link)
 
 
-@pytest.mark.skipif(can_not_symlink(), reason="symlink function not available")
 def test_remove_link_to_dir():
     with tempdir() as td:
         dst_link = join(td, "test_link")
         src_dir = join(td, "test_dir")
         _write_file(src_dir, "welcome to the ministry of silly walks")
-        os.symlink(src_dir, dst_link)
+        symlink(src_dir, dst_link)
         assert not islink(src_dir)
         assert islink(dst_link)
         assert rm_rf(dst_link)
@@ -96,6 +93,25 @@ def test_remove_link_to_dir():
         assert not isdir(src_dir)
         assert not islink(src_dir)
         assert not lexists(dst_link)
+
+
+def test_rm_rf_does_not_follow_symlinks():
+    with TemporaryDirectory() as tmp:
+        # make a file in some temp folder
+        real_file = os.path.join(tmp, 'testfile')
+        with open(real_file, 'w') as f:
+            f.write('weee')
+        # make a subfolder
+        subdir = os.path.join(tmp, 'subfolder')
+        os.makedirs(subdir)
+        # link to the file in the subfolder
+        link_path = join(subdir, 'file_link')
+        create_link(real_file, link_path, link_type=LinkType.softlink)
+        assert islink(link_path)
+        # rm_rf the subfolder
+        rm_rf(subdir)
+        # assert that the file still exists in the root folder
+        assert os.path.isfile(real_file)
 
 
 def test_move_to_trash():